moved conditional_skip_evaluation logic from target_dose.rb to a module conditional_skip_evaulation.rb
